LPS-179567 Extract CommercePriceListDetailsScreenNavigationCategory methods to a CommercePriceListDetailsScreenNavigationEntry component and extends CommercePriceListDetailsScreenNavigationCategory in it
